Restructure the pro space with a left sidebar and dedicated sections
Dashboard (tiles + next chantier + preview), Mes chantiers, Pointage
and Disponibilités become separate pages behind a ProNav sidebar,
mirroring the admin layout. Login/reset pages stay bare.
